requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.3
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
